Move ModTools into its own component.
